requirements.txt
Flask
Flask-SQLAlchemy
google-generativeai
numpy
pandas
sklearn
seaborn
matplotlib
plotly